refactor: extract shared SDF scaffolding for volumetric modifiers
Pull the generic core of the Voronoi-lamp mesh path into a reusable
sdfModifierMesh(mesh, { resolution, bandWorld, watertight,
smoothIterations }, combine) in src/surface/sdfModifier.ts. It owns
rasterize -> BVH narrow-band signed distance (signed by the closest face
normal) -> padded field sweep -> keepLargestFaceConnected ->
interpolated surfaceNetsField -> largestMeshComponent -> Taubin.

A feature now only supplies a combine({ d, x, y, z, voxelSize }) => number
closure (< 0 = inside). Re-point voronoiLampSdf at it (now ~75 lines,
max(shell, strut)); output is identical (unit + e2e lamp tests unchanged).

Foundation for upcoming volumetric ops (hollow/vase, gyroid/honeycomb
infill, conformal lattice, wireframe cage, engrave-through) so each is an
additive combine + thin registration rather than a copy of the pipeline.

No user-facing behavior change.
